+++ frontend/src/App.tsx
@@ -1,40 +1,19 @@
 import './App.css';
-import {useEffect, useState} from 'react';
-import axios from 'axios';
 import HobbyList from './components/HobbyList';
-import {Hobby, HobbyWithoutID} from './models';
 import Header from './components/Header';
 import Button from '@mui/material/Button';
 import {Route, Routes, useNavigate} from 'react-router-dom';
 import AddForm from './components/AddForm';
+import useHobbies from "./hooks/useHobbies.ts";
 
 
 export default function App() {
-    const [hobbies, setHobbies] = useState<Hobby[]>([]);
-    const colors = [' ', 'lightblue', 'lightgreen', 'pink', 'violet', 'orange', 'turquoise'];
 
     const navigate = useNavigate();
+    const colors = [' ', 'lightblue', 'lightgreen', 'pink', 'violet', 'orange', 'turquoise'];
 
-    useEffect(() => {
-        axios
-            .get('api/hobbies')
-            .then((response) => response.data)
-            .catch(console.error)
-            .then((data) => setHobbies(data));
-    }, []);
+    const {hobbies, handleAddHobby} = useHobbies()
 
-    function handleAddHobby(data: HobbyWithoutID) {
-        axios
-            .post('api/hobbies', data)
-            .then((response) => response.data)
-            .catch((error) => {
-                console.error(error);
-            })
-            .then((data) => {
-                setHobbies(data);
-            });
-        navigate('/');
-    }
 
     return (
         <main>
